Develop

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const Product = require('../models/product')
 const { validationResult } = require('express-validator/check')
+const fileHelper = require('../util/file')
 exports.getAddProduct = (req, res, next) => {
   res.render('admin/edit-product', {
     pageTitle: 'Add Product',
@@ -14,11 +15,25 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title
-  const imageUrl = req.body.imageUrl
+  const image = req.file
   const price = req.body.price
   const description = req.body.description
   const errors = validationResult(req)
-
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/edit-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
+        price: price,
+        description: description
+      },
+      errorMessage: 'Attached file is not an image.',
+      validationErrors: errors.array()
+    })
+  }
   if (!errors.isEmpty()) {
     return res.status(422).render('admin/edit-product', {
       pageTitle: 'Add Product',
@@ -27,7 +42,6 @@ exports.postAddProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: title,
-        imageUrl: imageUrl,
         price: price,
         description: description
       },
@@ -35,6 +49,8 @@ exports.postAddProduct = (req, res, next) => {
       validationErrors: errors.array()
     })
   }
+
+  const imageUrl = image.path
   const product = new Product({
     title: title,
     imageUrl: imageUrl,
@@ -85,7 +101,7 @@ exports.getEditProduct = (req, res, next) => {
 exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId
   const updatedTile = req.body.title
-  const updatedImageUrl = req.body.imageUrl
+  const image = req.file
   const updatedPrice = req.body.price
   const updatedDescription = req.body.description
 
@@ -99,7 +115,6 @@ exports.postEditProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: updatedTile,
-        imageUrl: updatedImageUrl,
         price: updatedPrice,
         description: updatedDescription,
         _id: prodId
@@ -116,7 +131,10 @@ exports.postEditProduct = (req, res, next) => {
       }
       product.title = updatedTile
       product.price = updatedPrice
-      product.imageUrl = updatedImageUrl
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl)
+        product.imageUrl = image.path
+      }
       product.description = updatedDescription
       return product.save().then(result => {
         res.redirect('/admin/products')
@@ -147,7 +165,13 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const productId = req.body.productId
-  Product.deleteOne({ _id: productId, userId: req.user })
+  Product.findById(productId).then(product => {
+    if (!product) {
+      return next(new Error('Product not found.'))
+    }
+    fileHelper.deleteFile(product.imageUrl)
+    return Product.deleteOne({ _id: productId, userId: req.user })
+  })
     .then(() => {
       res.redirect('/admin/products')
     })
